requirements.txt
discord.py
audioop-lts; python_version >= "3.13"